refactor: Moved core of event handling to the new module event-system.
