+++ src/app/core/services/basic-rundown-state.service.ts
@@ -1,10 +1,10 @@
 import { Injectable, OnDestroy } from '@angular/core'
 import { BehaviorSubject, lastValueFrom, SubscriptionLike } from 'rxjs'
 import { RundownEventObserver } from './events/rundown-event-observer.service'
-import { Unsubscribe } from './events/event-observer.interface'
-import { ConnectionStatusObserver } from './events/connection-status-observer.service'
+import { Unsubscribe } from '../../event-system/services/event-observer.interface'
+import { ConnectionStatusObserver } from '../../event-system/services/connection-status-observer.service'
 import { BasicRundown } from '../models/basic-rundown'
-import { BasicRundownService } from '../interfaces/basic-rundown-service'
+import { BasicRundownService } from '../interfaces/basic-rundown-service.interface'
 
 @Injectable()
 export class BasicRundownStateService implements OnDestroy {
