+++ src/app/event-system/services/connection-status-observer.service.ts
