+++ src/app/event-system/services/event-observer.interface.ts
