+++ src/app/event-system/services/robust-websocket.factory.ts
