+++ src/app/event-system/services/websocket-event-observer.service.ts
